restructure Repository, extract alert into single util function
