+++ src/components/AppBar/SignOut.jsx
@@ -1,7 +1,8 @@
-import { View, StyleSheet, Pressable, Alert } from "react-native"
-import Text from "../Text"
-import useSignOut from "../../hooks/useSignOut"
-import { useNavigate } from "react-router-native"
+import { View, StyleSheet, Pressable } from "react-native";
+import Text from "../Text";
+import useSignOut from "../../hooks/useSignOut";
+import { useNavigate } from "react-router-native";
+import { alert } from "../../utils/alert";
 
 const styles = StyleSheet.create({
   tab: {
@@ -10,38 +11,36 @@ const styles = StyleSheet.create({
     alignItems: "center",
     padding: 15,
   },
-})
+});
 
 const SignOut = () => {
-  const [signOut] = useSignOut()
-  const navigate = useNavigate()
+  const [signOut] = useSignOut();
+  const navigate = useNavigate();
 
-  const onSignOutPressed = () => {
-    Alert.alert("Log Out", "Are you sure you want to log out", [
-      // The "Yes" button
-      {
-        text: "Yes",
-        onPress: async () => {
-          await signOut()
-          navigate("/SignIn")
-        },
-      },
-      // The "No" button
-      {
-        text: "No",
-      },
-    ])
-  }
+  const onSignOutPressed = async () => {
+    await signOut();
+    navigate("/SignIn");
+  };
+
+  const onSignOut = () => {
+    alert(
+      "Log out",
+      "Are you sure you want to log out",
+      "Yes",
+      "No",
+      onSignOutPressed
+    );
+  };
 
   return (
     <View style={styles.tab}>
-      <Pressable onPress={onSignOutPressed}>
+      <Pressable onPress={onSignOut}>
         <Text color="barTab" fontWeight="bold">
           SignOut
         </Text>
       </Pressable>
     </View>
-  )
-}
+  );
+};
 
-export default SignOut
+export default SignOut;
